+++ Microsoft.DotNet.Interactive.Tests/DirectiveTests.cs
@@ -9,7 +9,6 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.DotNet.Interactive.CSharp;
 using Microsoft.DotNet.Interactive.Events;
-using WorkspaceServer.Kernel;
 using Xunit;
 
 namespace Microsoft.DotNet.Interactive.Tests
