+++ Microsoft.DotNet.Interactive.Tests/KernelClientTests.Kernel_can_be_interacted_using_kernel_client.approved.json
