+++ Microsoft.DotNet.Interactive.Tests/KernelClientTests.Kernel_client_surfaces_json_errors.approved.json
