+++ Microsoft.DotNet.Interactive.Tests/Microsoft.DotNet.Interactive.Tests.csproj
@@ -53,7 +53,4 @@
   <ItemGroup>
     <PackageReference Update="Microsoft.NET.Test.Sdk" Version="16.2.0" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Kernel\" />
-  </ItemGroup>
 </Project>
