+++ Microsoft.DotNet.Interactive.Tests/StreamKernelCommandTests.Serialization_contract_is_not_broken.approved.json
