+++ Microsoft.DotNet.Interactive.Tests/StreamKernelCommandTests.cs
@@ -3,13 +3,12 @@
 
 using System.Diagnostics;
 using Assent;
-using Microsoft.DotNet.Interactive;
 using Microsoft.DotNet.Interactive.Commands;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Xunit;
 
-namespace WorkspaceServer.Tests.Kernel
+namespace Microsoft.DotNet.Interactive.Tests
 {
     public class StreamKernelCommandTests
     {
