+++ dotnet-interactive/Startup.cs
@@ -3,28 +3,16 @@
 
 using System;
 using System.Linq;
-using System.Net.Mime;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.HttpOverrides;
-using Microsoft.AspNetCore.ResponseCompression;
 using Microsoft.DotNet.Interactive.App.CommandLine;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using MLS.Agent.CommandLine;
-using MLS.Agent.Tools;
-using Pocket;
-using static Pocket.Logger<Microsoft.DotNet.Interactive.App.Startup>;
 
 namespace Microsoft.DotNet.Interactive.App
 {
     public class Startup
     {
-        private readonly CompositeDisposable _disposables = new CompositeDisposable
-        {
-            () => Logger<Program>.Log.Event("AgentStopping")
-        };
-
         public Startup(
             IHostEnvironment env,
             StartupOptions startupOptions)
@@ -46,35 +34,12 @@ namespace Microsoft.DotNet.Interactive.App
 
         public void ConfigureServices(IServiceCollection services)
         {
-            using (var operation = Log.OnEnterAndConfirmOnExit())
-            {
-                services.AddSingleton(Configuration);
-
-                services.AddResponseCompression(options =>
-                {
-                    options.MimeTypes = ResponseCompressionDefaults.MimeTypes.Concat(new[]
-                    {
-                        MediaTypeNames.Application.Octet,
-                    });
-                });
-
-                services.Configure<ForwardedHeadersOptions>(options => { options.ForwardedHeaders = ForwardedHeaders.All; });
-
-                operation.Succeed();
-            }
         }
 
         public void Configure(
             IApplicationBuilder app,
-            IHostApplicationLifetime lifetime,
-            IDirectoryAccessor directoryAccessor)
+            IHostApplicationLifetime lifetime)
         {
-            using (var operation = Log.OnEnterAndConfirmOnExit())
-            {
-                lifetime.ApplicationStopping.Register(() => _disposables.Dispose());
-
-                operation.Succeed();
-            }
         }
     }
 }
